+++ lib/src/media_container.dart
@@ -23,6 +23,7 @@ class MediaContainer extends StatefulWidget {
     bool? autoPlayVideo = false,
     bool? enableVideoFullscreen = true,
     this.videoControlBarBgColor,
+    this.videoPlaceholder,
   })  : autoPlayVideo = autoPlayVideo ?? false,
         enableVideoFullscreen = enableVideoFullscreen ?? true,
         super(key: key);
@@ -34,6 +35,7 @@ class MediaContainer extends StatefulWidget {
   final bool autoPlayVideo;
   final bool enableVideoFullscreen;
   final Color? videoControlBarBgColor;
+  final Widget? videoPlaceholder;
 
   @override
   _MediaContainerState createState() => _MediaContainerState();
@@ -62,6 +64,7 @@ class _MediaContainerState extends State<MediaContainer>
         autoPlay: widget.autoPlayVideo,
         enableFullscreen: widget.enableVideoFullscreen,
         controlBarColor: widget.videoControlBarBgColor,
+        videoPlaceholder: widget.videoPlaceholder,
       );
     } else if (widget.tweetVM.getDisplayTweet().hasPhoto) {
       switch (widget.tweetVM.getDisplayTweet().allPhotos.length) {